refactor: improve test helper file names
