+++ src/index.test.ts
@@ -1,9 +1,9 @@
 import { FetchMock } from 'jest-fetch-mock';
 import 'jest-localstorage-mock';
-import * as data from './tests/example-data.json';
+import * as data from './test/testdata.json';
 import IStorageProvider from './storage-provider';
 import { EVENTS, IConfig, IMutableContext, UnleashClient } from './index';
-import { getTypeSafeRequest, getTypeSafeRequestUrl } from './tests/util';
+import { getTypeSafeRequest, getTypeSafeRequestUrl } from './test';
 
 jest.useFakeTimers();
